fix: `ObjectAlignment` enum names (#74)
Name of some of the enum values of ObjectAlignment were not matching the values from Tiled. As a result, tilesets with topLeft, topRight, bottomLeft and bottomRight alignment couldn't be parsed. https://doc.mapeditor.org/en/stable/reference/tmx-map-format/#tmx-tileset

image

Another problem that was discovered after fixing the names was with the byName method from the EnumByName extension. That method was not using names from ObjectAlignmentExtension that Tiled defines. To fix that, this PR adds a byName static method for ObjectAlignment which uses the correct names.